EZP-31357: Universal Discovery improvements
