+++ src/bundle/Controller/UniversalDiscoveryController.php
@@ -8,95 +8,33 @@ declare(strict_types=1);
 
 namespace EzSystems\EzPlatformAdminUiBundle\Controller;
 
-use eZ\Publish\API\Repository\BookmarkService;
-use eZ\Publish\API\Repository\ContentService;
-use eZ\Publish\API\Repository\ContentTypeService;
-use eZ\Publish\API\Repository\LocationService;
-use eZ\Publish\API\Repository\Repository;
-use eZ\Publish\API\Repository\SearchService;
 use eZ\Publish\API\Repository\Values\Content\Location;
-use eZ\Publish\API\Repository\Values\Content\LocationQuery;
 use eZ\Publish\API\Repository\Values\Content\Query;
-use eZ\Publish\API\Repository\Values\Content\Search\SearchHit;
-use eZ\Publish\API\Repository\Values\User\Limitation;
-use EzSystems\EzPlatformAdminUi\Permission\LookupLimitationsTransformer;
-use EzSystems\EzPlatformAdminUi\Permission\PermissionCheckerInterface;
-use EzSystems\EzPlatformAdminUi\QueryType\LocationPathQueryType;
-use EzSystems\EzPlatformRest\Output\Visitor;
-use EzSystems\EzPlatformRest\Server\Values\Version;
+use EzSystems\EzPlatformAdminUi\UniversalDiscovery\Provider;
 use Symfony\Component\HttpFoundation\JsonResponse;
 use Symfony\Component\HttpFoundation\Request;
 
 class UniversalDiscoveryController extends Controller
 {
-    private const ROOT_LOCATION_ID = 1;
-    private const COLUMNS_NUMBER = 4;
-
-    private const SORT_CLAUSE_DATE_PUBLISHED = 'DatePublished';
-    private const SORT_CLAUSE_CONTENT_NAME = 'ContentName';
-
-    /** @var \eZ\Publish\API\Repository\Repository */
-    private $repository;
-
-    /** @var \eZ\Publish\API\Repository\LocationService */
-    private $locationService;
-
-    /** @var \eZ\Publish\API\Repository\ContentTypeService */
-    private $contentTypeService;
-
-    /** @var \eZ\Publish\API\Repository\SearchService */
-    private $searchService;
-
-    /** @var \EzSystems\EzPlatformRest\Output\Visitor */
-    private $visitor;
-
-    /** @var \eZ\Publish\API\Repository\BookmarkService */
-    private $bookmarkService;
-
-    /** @var \eZ\Publish\API\Repository\ContentService */
-    private $contentService;
-
-    /** @var \EzSystems\EzPlatformAdminUi\Permission\PermissionCheckerInterface */
-    private $permissionChecker;
-
-    /** @var \EzSystems\EzPlatformAdminUi\Permission\LookupLimitationsTransformer */
-    private $lookupLimitationsTransformer;
-
-    /** @var \EzSystems\EzPlatformAdminUi\QueryType\LocationPathQueryType */
-    private $locationPathQueryType;
-
-    private $sortClauseClassMap = [
-        self::SORT_CLAUSE_DATE_PUBLISHED => Query\SortClause\DatePublished::class,
-        self::SORT_CLAUSE_CONTENT_NAME => Query\SortClause\ContentName::class,
-    ];
-
-    private $availableSortOrder = [
-        Query::SORT_ASC,
-        Query::SORT_DESC,
-    ];
+    /** @var \EzSystems\EzPlatformAdminUi\UniversalDiscovery\Provider */
+    private $provider;
 
     public function __construct(
-        Repository $repository,
-        LocationService $locationService,
-        ContentTypeService $contentTypeService,
-        SearchService $searchService,
-        BookmarkService $bookmarkService,
-        ContentService $contentService,
-        Visitor $visitor,
-        PermissionCheckerInterface $permissionChecker,
-        LookupLimitationsTransformer $lookupLimitationsTransformer,
-        LocationPathQueryType $locationPathQueryType
+        Provider $provider
     ) {
-        $this->repository = $repository;
-        $this->searchService = $searchService;
-        $this->contentTypeService = $contentTypeService;
-        $this->locationService = $locationService;
-        $this->visitor = $visitor;
-        $this->bookmarkService = $bookmarkService;
-        $this->contentService = $contentService;
-        $this->permissionChecker = $permissionChecker;
-        $this->lookupLimitationsTransformer = $lookupLimitationsTransformer;
-        $this->locationPathQueryType = $locationPathQueryType;
+        $this->provider = $provider;
+    }
+
+    public function locationsAction(Request $request)
+    {
+        return new JsonResponse(
+            $this->provider->getLocations(
+                explode(
+                    ',',
+                    $request->query->get('locationIds', '')
+                )
+            )
+        );
     }
 
     public function locationAction(
@@ -105,13 +43,13 @@ class UniversalDiscoveryController extends Controller
     ): JsonResponse {
         $offset = $request->query->getInt('offset', 0);
         $limit = $request->query->getInt('limit', 25);
-        $sortClauseName = $request->query->getAlpha('sortClause', self::SORT_CLAUSE_DATE_PUBLISHED);
+        $sortClauseName = $request->query->getAlpha('sortClause', Provider::SORT_CLAUSE_DATE_PUBLISHED);
         $sortOrder = $request->query->getAlpha('sortOrder', Query::SORT_ASC);
 
-        $sortClause = $this->getSortClause($sortClauseName, $sortOrder);
+        $sortClause = $this->provider->getSortClause($sortClauseName, $sortOrder);
 
         return new JsonResponse(
-            $this->getLocationData($locationId, $offset, $limit, $sortClause)
+            $this->provider->getLocationData($locationId, $offset, $limit, $sortClause)
         );
     }
 
@@ -121,13 +59,13 @@ class UniversalDiscoveryController extends Controller
     ): JsonResponse {
         $offset = $request->query->getInt('offset', 0);
         $limit = $request->query->getInt('limit', 25);
-        $sortClauseName = $request->query->getAlpha('sortClause', self::SORT_CLAUSE_DATE_PUBLISHED);
+        $sortClauseName = $request->query->getAlpha('sortClause', Provider::SORT_CLAUSE_DATE_PUBLISHED);
         $sortOrder = $request->query->getAlpha('sortOrder', Query::SORT_ASC);
 
-        $sortClause = $this->getSortClause($sortClauseName, $sortOrder);
+        $sortClause = $this->provider->getSortClause($sortClauseName, $sortOrder);
 
         return new JsonResponse(
-            $this->getLocationGridViewData($locationId, $offset, $limit, $sortClause)
+            $this->provider->getLocationGridViewData($locationId, $offset, $limit, $sortClause)
         );
     }
 
@@ -136,19 +74,19 @@ class UniversalDiscoveryController extends Controller
         int $locationId
     ): JsonResponse {
         $limit = $request->query->getInt('limit', 25);
-        $sortClauseName = $request->query->getAlpha('sortClause', self::SORT_CLAUSE_DATE_PUBLISHED);
+        $sortClauseName = $request->query->getAlpha('sortClause', Provider::SORT_CLAUSE_DATE_PUBLISHED);
         $sortOrder = $request->query->getAlpha('sortOrder', Query::SORT_ASC);
-        $rootLocationId = $request->query->getInt('rootLocationId', self::ROOT_LOCATION_ID);
+        $rootLocationId = $request->query->getInt('rootLocationId', Provider::ROOT_LOCATION_ID);
 
-        $sortClause = $this->getSortClause($sortClauseName, $sortOrder);
-        $breadcrumbLocations = $this->getBreadcrumbLocations($locationId);
+        $sortClause = $this->provider->getSortClause($sortClauseName, $sortOrder);
+        $breadcrumbLocations = $this->provider->getBreadcrumbLocations($locationId);
 
-        $columns = $this->getColumns($locationId, $limit, $sortClause, false, $rootLocationId);
+        $columns = $this->provider->getColumns($locationId, $limit, $sortClause, false, $rootLocationId);
 
         return new JsonResponse([
             'breadcrumb' => array_map(
                 function (Location $location) {
-                    return $this->getRestFormat($location);
+                    return $this->provider->getRestFormat($location);
                 },
                 $breadcrumbLocations
             ),
@@ -161,250 +99,22 @@ class UniversalDiscoveryController extends Controller
         int $locationId
     ): JsonResponse {
         $limit = $request->query->getInt('limit', 25);
-        $sortClauseName = $request->query->getAlpha('sortClause', self::SORT_CLAUSE_DATE_PUBLISHED);
+        $sortClauseName = $request->query->getAlpha('sortClause', Provider::SORT_CLAUSE_DATE_PUBLISHED);
         $sortOrder = $request->query->getAlpha('sortOrder', Query::SORT_ASC);
-        $rootLocationId = $request->query->getInt('rootLocationId', self::ROOT_LOCATION_ID);
+        $rootLocationId = $request->query->getInt('rootLocationId', Provider::ROOT_LOCATION_ID);
 
-        $sortClause = $this->getSortClause($sortClauseName, $sortOrder);
+        $sortClause = $this->provider->getSortClause($sortClauseName, $sortOrder);
 
-        $columns = $this->getColumns($locationId, $limit, $sortClause, true, $rootLocationId);
+        $columns = $this->provider->getColumns($locationId, $limit, $sortClause, true, $rootLocationId);
 
         return new JsonResponse([
             'breadcrumb' => array_map(
                 function (Location $location) {
-                    return $this->getRestFormat($location);
+                    return $this->provider->getRestFormat($location);
                 },
-                $this->getBreadcrumbLocations($locationId)
+                $this->provider->getBreadcrumbLocations($locationId)
             ),
             'columns' => $columns,
         ]);
-    }
-
-    private function getColumns(
-        int $locationId,
-        int $limit,
-        Query\SortClause $sortClause,
-        bool $gridView = false,
-        int $rootLocationId = self::ROOT_LOCATION_ID
-    ): array {
-        $location = $this->locationService->loadLocation($locationId);
-
-        $locationPath = $this->getRelativeLocationPath($rootLocationId, $location->path);
-        $locationPathCount = count($locationPath);
-
-        if ($locationPathCount > self::COLUMNS_NUMBER) {
-            $locationPathLast = $locationPathCount - 1;
-
-            $columnLocations = [
-                $locationPath[0], // First
-                $locationPath[$locationPathLast - 2],
-                $locationPath[$locationPathLast - 1],
-                $locationPath[$locationPathLast], // Last
-            ];
-        } else {
-            $columnLocations = $locationPath;
-        }
-
-        $columns = [];
-        foreach ($columnLocations as $columnLocationId) {
-            $columnLocationId = (int)$columnLocationId;
-            $columnLocation = ($columnLocationId !== self::ROOT_LOCATION_ID)
-                ? $this->getRestFormat($this->locationService->loadLocation($columnLocationId))
-                : null;
-
-            $columns[$columnLocationId] = [
-                'location' => $columnLocation,
-                'subitems' => $this->getSubitemLocations($columnLocationId, 0, $limit, $sortClause),
-            ];
-        }
-
-        $columns[$locationId] = $gridView
-            ? $this->getLocationGridViewData($locationId, 0, $limit, $sortClause)
-            : $this->getLocationData($locationId, 0, $limit, $sortClause);
-
-        return $columns;
-    }
-
-    private function getBreadcrumbLocations(
-        int $locationId,
-        int $rootLocationId = self::ROOT_LOCATION_ID
-    ): array {
-        $searchResult = $this->searchService->findLocations(
-            $this->locationPathQueryType->getQuery([
-                'location' => $this->locationService->loadLocation($locationId),
-                'rootLocationId' => $rootLocationId,
-            ])
-        );
-
-        return array_map(
-            function (SearchHit $searchHit) {
-                return $searchHit->valueObject;
-            }, $searchResult->searchHits
-        );
-    }
-
-    private function getLocationPermissionRestrictions(Location $location): array
-    {
-        $lookupLimitationsResult = $this->permissionChecker->getContentCreateLimitations($location);
-        $limitationsValues = $this->lookupLimitationsTransformer->getGroupedLimitationValues(
-            $lookupLimitationsResult,
-            [Limitation::CONTENTTYPE, Limitation::LANGUAGE]
-        );
-
-        return [
-            'hasAccess' => $lookupLimitationsResult->hasAccess,
-            'restrictedContentTypeIds' => $limitationsValues[Limitation::CONTENTTYPE],
-            'restrictedLanguageCodes' => $limitationsValues[Limitation::LANGUAGE],
-        ];
-    }
-
-    private function getSubitemContents(
-        int $locationId,
-        int $offset,
-        int $limit,
-        Query\SortClause $sortClause
-    ): array {
-        $searchResult = $this->searchService->findContent(
-            new LocationQuery([
-                'filter' => new Query\Criterion\ParentLocationId($locationId),
-                'sortClauses' => [$sortClause],
-                'offset' => $offset,
-                'limit' => $limit,
-            ])
-        );
-
-        return array_map(
-            function (SearchHit $searchHit) {
-                /** @var \eZ\Publish\API\Repository\Values\Content\Content $content */
-                $content = $searchHit->valueObject;
-
-                return $this->getRestFormat(
-                    new Version($content, $content->getContentType(), [])
-                );
-            },
-            $searchResult->searchHits
-        );
-    }
-
-    private function getSubitemLocations(
-        int $locationId,
-        int $offset,
-        int $limit,
-        Query\SortClause $sortClause
-    ): array {
-        $searchResult = $this->searchService->findLocations(
-            new LocationQuery([
-                'filter' => new Query\Criterion\ParentLocationId($locationId),
-                'sortClauses' => [$sortClause],
-                'offset' => $offset,
-                'limit' => $limit,
-            ])
-        );
-
-        return [
-            'locations' => array_map(
-                function (SearchHit $searchHit) {
-                    return $this->getRestFormat($searchHit->valueObject);
-                },
-                $searchResult->searchHits
-            ),
-            'totalCount' => $searchResult->totalCount,
-        ];
-    }
-
-    private function getLocationData(
-        int $locationId,
-        int $offset,
-        int $limit,
-        Query\SortClause $sortClause
-    ): array {
-        if ($locationId === self::ROOT_LOCATION_ID) {
-            return [
-                'subitems' => $this->getSubitemLocations($locationId, $offset, $limit, $sortClause),
-            ];
-        }
-
-        $location = $this->locationService->loadLocation($locationId);
-        $content = $this->contentService->loadContentByContentInfo($location->getContentInfo());
-        $contentType = $this->contentTypeService->loadContentType($location->getContentInfo()->contentTypeId);
-
-        return [
-            'location' => $this->getRestFormat($location),
-            'bookmarked' => $this->bookmarkService->isBookmarked($location),
-            'permissions' => $this->getLocationPermissionRestrictions($location),
-            'version' => $this->getRestFormat(new Version($content, $contentType, [])),
-            'subitems' => $this->getSubitemLocations($locationId, $offset, $limit, $sortClause),
-        ];
-    }
-
-    private function getLocationGridViewData(
-        int $locationId,
-        int $offset,
-        int $limit,
-        Query\SortClause $sortClause
-    ): array {
-        if ($locationId === self::ROOT_LOCATION_ID) {
-            $locations = $this->getSubitemLocations($locationId, $offset, $limit, $sortClause);
-            $versions = $this->getSubitemContents($locationId, $offset, $limit, $sortClause);
-
-            return [
-                'subitems' => [
-                    'locations' => $locations['locations'],
-                    'totalCount' => $locations['totalCount'],
-                    'versions' => $versions,
-                ],
-            ];
-        }
-
-        $location = $this->locationService->loadLocation($locationId);
-        $content = $this->contentService->loadContentByContentInfo($location->getContentInfo());
-        $contentType = $this->contentTypeService->loadContentType($location->getContentInfo()->contentTypeId);
-
-        $locations = $this->getSubitemLocations($locationId, $offset, $limit, $sortClause);
-        $versions = $this->getSubitemContents($locationId, $offset, $limit, $sortClause);
-
-        return [
-            'location' => $this->getRestFormat($location),
-            'bookmarked' => $this->bookmarkService->isBookmarked($location),
-            'permissions' => $this->getLocationPermissionRestrictions($location),
-            'version' => $this->getRestFormat(new Version($content, $contentType, [])),
-            'subitems' => [
-                'locations' => $locations['locations'],
-                'totalCount' => $locations['totalCount'],
-                'versions' => $versions,
-            ],
-        ];
-    }
-
-    private function getRestFormat($valueObject): array
-    {
-        return json_decode(
-            $this->visitor->visit($valueObject)->getContent(),
-            true
-        );
-    }
-
-    private function getSortClause(string $sortClauseName, string $sortOrder): Query\SortClause
-    {
-        $sortClauseClass = $this->sortClauseClassMap[$sortClauseName] ?? $this->sortClauseClassMap[self::SORT_CLAUSE_DATE_PUBLISHED];
-        $sortOrder = !in_array($sortOrder, $this->availableSortOrder)
-            ? $this->availableSortOrder[0]
-            : $sortOrder;
-
-        return new $sortClauseClass($sortOrder);
-    }
-
-    private function getRelativeLocationPath(int $locationId, array $locationPath): array
-    {
-        $locationIds = array_values($locationPath);
-
-        $index = array_search($locationId, $locationIds);
-
-        // Location is not part of path
-        if ($index === false) {
-            return [];
-        }
-
-        return array_slice($locationIds, $index);
     }
 }
