+++ src/lib/Permission/PermissionChecker.php
@@ -83,7 +83,7 @@ class PermissionChecker implements PermissionCheckerInterface
      *
      * @return array
      */
-    public function getRestrictions($hasAccess, string $class): array
+    public function getRestrictions(array $hasAccess, string $class): array
     {
         $restrictions = [];
         $oneOfPoliciesHasNoLimitation = false;
@@ -190,9 +190,6 @@ class PermissionChecker implements PermissionCheckerInterface
         return true;
     }
 
-    /**
-     * {@inheritdoc}
-     */
     public function getContentCreateLimitations(Location $parentLocation): LookupLimitationResult
     {
         $contentType = $this->contentTypeService->loadContentType($parentLocation->contentInfo->contentTypeId);
@@ -207,6 +204,21 @@ class PermissionChecker implements PermissionCheckerInterface
         return $this->permissionResolver->lookupLimitations('content', 'create',
             $contentCreateStruct,
             [$versionBuilder->build(), $locationCreateStruct],
+            [Limitation::CONTENTTYPE, Limitation::LANGUAGE]
+        );
+    }
+
+    public function getContentUpdateLimitations(Location $location): LookupLimitationResult
+    {
+        $contentUpdateStruct = $this->contentService->newContentUpdateStruct(); //$contentType, $parentLocation->contentInfo->mainLanguageCode);
+
+        $versionBuilder = new VersionBuilder();
+        $versionBuilder->translateToAnyLanguageOf($this->getActiveLanguageCodes());
+        $versionBuilder->createFromAnyContentTypeOf($this->getContentTypeIds());
+
+        return $this->permissionResolver->lookupLimitations('content', 'edit',
+            $contentUpdateStruct,
+            [$versionBuilder->build(), $location],
             [Limitation::CONTENTTYPE, Limitation::LANGUAGE]
         );
     }
