+++ src/lib/Permission/PermissionCheckerInterface.php
@@ -11,32 +11,28 @@ use eZ\Publish\API\Repository\Values\User\LookupLimitationResult;
 
 interface PermissionCheckerInterface
 {
-    /**
-     * @param $hasAccess
-     * @param string $class
-     *
-     * @return array
-     */
-    public function getRestrictions($hasAccess, string $class): array;
+    public function getRestrictions(array $hasAccess, string $class): array;
 
     /**
-     * @param \eZ\Publish\API\Repository\Values\Content\Location $location
      * @param array|bool $hasAccess
-     *
-     * @return bool
      */
     public function canCreateInLocation(Location $location, $hasAccess): bool;
 
     /**
      * @internal
      *
-     * @param \eZ\Publish\API\Repository\Values\Content\Location $parentLocation
-     *
-     * @return \eZ\Publish\API\Repository\Values\User\LookupLimitationResult
-     *
      * @throws \eZ\Publish\API\Repository\Exceptions\BadStateException
      * @throws \eZ\Publish\API\Repository\Exceptions\InvalidArgumentException
      * @throws \eZ\Publish\API\Repository\Exceptions\NotFoundException
      */
     public function getContentCreateLimitations(Location $parentLocation): LookupLimitationResult;
+
+    /**
+     * @internal
+     *
+     * @throws \eZ\Publish\API\Repository\Exceptions\BadStateException
+     * @throws \eZ\Publish\API\Repository\Exceptions\InvalidArgumentException
+     * @throws \eZ\Publish\API\Repository\Exceptions\NotFoundException
+     */
+    public function getContentUpdateLimitations(Location $parentLocation): LookupLimitationResult;
 }
